// ui/src/lib/generated/GreeterInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SessionInfo } from "./SessionInfo";

/**
 * Reply to `greeter.info`.
 */
export type GreeterInfo = { hostname: string, sessions: Array<SessionInfo>, last_user: string | null, 
/**
 * `SessionInfo::id` of the last session started from this greeter.
 */
last_session: string | null, };